+++ cypress/e2e/pyplot/spec.cy.ts
@@ -1,0 +1,14 @@
+describe("pyplot", () => {
+  before(() => {
+    cy.intercept("/project/settings").as("settings");
+    cy.visit("http://127.0.0.1:8000");
+    cy.wait(["@settings"]);
+  });
+
+  it("should be able to display an inline chart", () => {
+    cy.get("#welcome-screen").should("exist");
+
+    cy.get(".message").should("have.length", 1);
+    cy.get(".message").eq(0).find(".inline-image").should("have.length", 1);
+  });
+});