+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-Nothing unreleased yet.
+### Added
+
+- Pyplot chart element
 
 ## [0.3.0] - 2023-06-13